requirements.txt
# Core Framework and Dependencies
fastapi==0.104.1
starlette==0.27.0
uvicorn==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6
fastapi-cache2==0.2.2

# Database and ORM
alembic==1.12.1
psycopg2-binary==2.9.10
SQLAlchemy==2.0.41
greenlet==3.2.3

# Authentication and Security
bcrypt==4.0.1
cryptography==45.0.5
ecdsa==0.19.1
passlib==1.7.4
PyJWT==2.10.1
python-jose==3.3.0
pyasn1==0.6.1
rsa==4.9.1

# Background Tasks and Message Queues
aioredis==1.3.1
amqp==5.3.1
billiard==4.2.1
celery==5.3.4
hiredis==3.2.1
kombu==5.5.4
redis==5.0.1
vine==5.1.0

# Data Processing and Analysis
contourpy==1.3.2
matplotlib==3.10.3
networkx==3.5
numpy==2.3.1
pandas==2.3.1
scipy==1.16.1

# Excel and Document Processing
et_xmlfile==2.0.0
openpyxl==3.1.2
XlsxWriter==3.1.9

# PDF and Document Generation
pydyf==0.11.0
reportlab==4.0.8
weasyprint==61.2
cssselect2==0.8.0
html5lib==1.1
tinycss2==1.4.0
webencodings==0.5.1
zopfli==0.2.3.post1

# Image Processing
pillow==11.3.0

# HTTP and Networking
certifi==2025.7.14
charset-normalizer==3.4.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
idna==3.10
requests==2.32.4
urllib3==2.5.0
wsproto==1.2.0

# WebSocket Support
python-engineio==4.12.2
python-socketio==5.10.0
simple-websocket==1.1.0
websockets==12.0

# Cloud Storage
azure-core==1.35.0
azure-storage-blob==12.19.0
boto3==1.34.34
botocore==1.34.162
jmespath==1.0.1
s3transfer==0.10.4

# Validation and Configuration
email_validator==2.2.0
pydantic==2.11.7
pydantic-settings==2.0.3
pydantic_core==2.33.2
python-dotenv==1.1.1

# Date and Time
isodate==0.7.2
pendulum==3.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

# Development and Testing
black==23.11.0
coverage==7.9.2
execnet==2.1.1
factory_boy==3.3.3
Faker==37.4.2
flake8==6.1.0
iniconfig==2.1.0
isort==5.12.0
mccabe==0.7.0
pluggy==1.6.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==6.2.1
pytest-mock==3.14.1
pytest-xdist==3.8.0
pycodestyle==2.11.1
psutil==6.1.1
pyflakes==3.1.0

# Utilities and Helpers
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
bidict==0.23.1
Brotli==1.1.0
cffi==1.17.1
chardet==5.2.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cycler==0.12.1
dnspython==2.7.0
fonttools==4.59.0
hiredis==3.2.1
kiwisolver==1.4.8
Mako==1.3.10
MarkupSafe==3.0.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
pycparser==2.22
Pygments==2.19.2
pyparsing==3.2.3
pyphen==0.17.2
PyYAML==6.0.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1
watchfiles==1.1.0
wcwidth==0.2.13
wheel==0.45.1
pyotp>=2.9.0
qrcode[pil]==8.2
Authlib==1.6.1
webauthn==2.6.0
prometheus_client==0.21.1 